apis/views.py
from django.contrib.auth.models import User
from rest_framework.authtoken.views import ObtainAuthToken
from rest_framework.authtoken.models import Token
from rest_framework.response import Response
from rest_framework import viewsets, status
from rest_framework.views import APIView
from rest_framework.permissions import IsAuthenticated
from rest_framework.decorators import api_view
from .models import Post, UserFields
from .serializers import PostSerializer


class PostsViewSet(viewsets.ViewSet):

    def list(self, request):
        posts = Post.objects.all()
        serializer = PostSerializer(posts, many=True)
        return Response(serializer.data)

class UserPostsView(APIView):
    permission_classes = [IsAuthenticated]

    def get(self, request):
        user = request.user
        return Response({
            'message': f"OK - {user}"
        })
        


class LoginAuthToken(ObtainAuthToken):

    def post(self, request, *args, **kwargs):
        serializer = self.serializer_class(data=request.data,
                                           context={'request': request})
        serializer.is_valid(raise_exception=True)
        user = serializer.validated_data['user']
        token, created = Token.objects.get_or_create(user=user)
        return Response({
            'token': token.key,
            'user_id': user.pk,
            'email': user.email
        })


@api_view(['POST'])
def register_user(request):
    username = request.POST.get('username')
    if User.objects.get(username=username):
        return Response(status=status.HTTP_409_CONFLICT)
    password = request.POST.get('password')
    user = User(
        username=username,
        password=password
    )
    user.save()
    return Response({"status": "Succesfuly create user"})
