requirements.txt
asgiref==3.4.1
boto3==1.18.49
botocore==1.21.49
Django==3.2.7
django-storages==1.11.1
djangorestframework==3.12.4
djongo==1.3.6
dnspython==1.16.0
environs==9.3.3
jmespath==0.10.0
marshmallow==3.13.0
Pillow==8.3.2
pymongo==3.12.0
python-dateutil==2.8.2
python-dotenv==0.19.0
pytz==2021.1
s3transfer==0.5.0
six==1.16.0
sqlparse==0.2.4
urllib3==1.26.7
